refactor(organization): move http_client & link_formatter to a util module
